Convert createFamilyMembers to typescript
