+++ lib/town/_common.ts
@@ -1,10 +1,9 @@
-import { Family } from '../../src/NPCGeneration/Relationships/createFamilyMembers'
 import { MaterialType, MaterialTypes } from '../buildings/structureData'
 import { Building, BuildingRelationship } from '../buildings/_common'
 import { Faction } from '../faction/_common'
 import { Profession } from '../npc-generation/professions'
 import { GenderName, RaceName } from '../npc-generation/raceTraits'
-import { NPC, NpcRelationship } from '../npc-generation/_common'
+import { Family, NPC, NpcRelationship } from '../npc-generation/_common'
 import { Road } from './roads'
 import { Weather } from '../src/weather'
 import { townData, TownType, PoliticalIdeology, EconomicIdeology } from './townData'
